split global into multiple files
